requirements.txt
asgiref==3.5.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
environ==1.0
Pillow==9.0.1
pyotp==2.6.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
